feat: implement DeepSpeed ZeRO-3 training
- Add DeepSpeed configuration for ZeRO-3 optimization
- Support distributed training across 3 GPUs
- Implement model fine-tuning pipeline
- Add training scripts and configurations
- Test and verify training functionality

Technical details:
- ZeRO-3 parameter partitioning
- Multi-GPU distributed training setup
- Memory optimization for large language models

+++ web_demo/glm_edge_1.5b_chat.py
@@ -15,7 +15,7 @@ model = AutoModelForCausalLM.from_pretrained(
     BASE_MODEL_PATH, 
     trust_remote_code=True,
     device_map='auto'
-).half()
+)
 
 # 加载 LoRA 权重
 model = PeftModel.from_pretrained(model, LORA_PATH)
